migrationの変更

+++ db/migrate/20201030130046_create_hobby_maps.rb
@@ -1,7 +1,7 @@
 class CreateHobbyMaps < ActiveRecord::Migration[5.2]
   def change
     create_table :hobby_maps do |t|
-      t.references :post, type: :integer, foreign_key: true
+      t.references :post, type: :bigint, foreign_key: true
       t.references :hobby, type: :integer, foreign_key: true
 
       t.timestamps
